refactor(gfql): factor lean-combine DataFrameT helpers into own module (#1755)
Review feedback: the lean-combine helpers (_lean_combine_enabled, _is_unique_ids,
_lean_engine_ok, _lean_intersect_full, _lean_prefilter_right) are reusable,
engine-agnostic DataFrameT helpers — move them out of chain.py into a dedicated
graphistry/compute/chain_lean_combine.py (mirrors the gfql_fast_paths.py #1731
convention: one-directional import, no back-edge, imports only leaf modules).
chain.py imports them; call sites + monkeypatch-based engagement tests unchanged.
11 lean-combine tests pass; mypy clean.
